+++ src/service/getGherkinCompletionItems.ts
@@ -1,5 +1,10 @@
 import { walkGherkinDocument } from '@cucumber/gherkin-utils'
-import { CompletionItem, CompletionItemKind, InsertTextFormat } from 'vscode-languageserver-types'
+import {
+  CompletionItem,
+  CompletionItemKind,
+  InsertTextFormat,
+  Position,
+} from 'vscode-languageserver-types'
 
 import { parseGherkinDocument } from '../gherkin/parseGherkinDocument.js'
 import { Index } from '../index/index.js'
@@ -8,7 +13,7 @@ import { lspCompletionSnippet } from './snippet/lspCompletionSnippet.js'
 // https://microsoft.github.io/language-server-protocol/specifications/specification-3-17/#textDocument_completion
 export function getGherkinCompletionItems(
   gherkinSource: string,
-  line: number,
+  position: Position,
   index: Index
 ): readonly CompletionItem[] {
   const { gherkinDocument } = parseGherkinDocument(gherkinSource)
@@ -20,7 +25,7 @@ export function getGherkinCompletionItems(
   let endCharacter: number
   walkGherkinDocument(gherkinDocument, undefined, {
     step(step) {
-      if (step.location.line === line + 1 && step.location.column !== undefined) {
+      if (step.location.line === position.line + 1 && step.location.column !== undefined) {
         text = step.text
         startCharacter = step.location.column + step.keyword.length - 1
         endCharacter = startCharacter + text.length
@@ -29,7 +34,11 @@ export function getGherkinCompletionItems(
   })
   if (text === undefined) return []
   const suggestions = index(text)
-  return suggestions.map((suggestion) => {
+  // https://github.com/microsoft/language-server-protocol/issues/898#issuecomment-593968008
+  return suggestions.map((suggestion, i) => {
+    // The index has already sorted the syggestions by match score.
+    // We're moving suggestions that are from undefined steps to the bottom
+    const sortText = (suggestion.matched ? i + 1000 : i + 2000).toString()
     const item: CompletionItem = {
       label: suggestion.label,
       insertTextFormat: InsertTextFormat.Snippet,
@@ -37,20 +46,28 @@ export function getGherkinCompletionItems(
       labelDetails: {
         ...(suggestion.matched ? {} : { detail: ' (undefined step)' }),
       },
+      // VSCode will only display suggestions that literally match the label.
+      // We're overriding this behaviour by setting filterText to what the user has typed,
+      // so that the suggestions are always displayed
+      filterText: text,
+      sortText,
       textEdit: {
         newText: lspCompletionSnippet(suggestion.segments),
         range: {
           start: {
-            line,
+            line: position.line,
             character: startCharacter,
           },
           end: {
-            line,
+            line: position.line,
             character: endCharacter,
           },
         },
       },
     }
+    const { label } = item
+    console.log({ label, sortText })
+    console.log()
     return item
   })
 }