+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.18.0] - 2022-05-10
+### Changed
+- Autocomlete suggestions now set `filterText` to prevent VSCode from filtering out suggestions
+- Autocomlete suggestions now set `sortText` to display suggestions from undefined steps last
+
 ## [0.17.0] - 2022-05-09
 ### Added
 - Ruby support ([#44](https://github.com/cucumber/language-service/pull/44))
@@ -141,7 +146,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ([#1732](https://github.com/cucumber/common/pull/1732)
 [aslakhellesoy](https://github.com/aslakhellesoy))
 
-[Unreleased]: https://github.com/cucumber/language-service/compare/v0.17.0...HEAD
+[Unreleased]: https://github.com/cucumber/language-service/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/cucumber/language-service/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/cucumber/language-service/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/cucumber/language-service/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/cucumber/language-service/compare/v0.14.4...v0.15.0